# doc/rest/ui/swagger.json
{
  "swagger": "2.0",
  "info" : {
    "description" : "<h1>Ant Media Server<\/h1><p>Ant Media Server supports RTMP, RTSP, MP4, HLS and WebRTC based on Red5<\/p>",
    "license" : {
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.txt",
      "name" : "Apache 2"
    },
    "version" : "1.2.0-SNAPSHOT",
    "title" : "Ant Media Server"
  },
  "basePath" : "/rest",
  "schemes" : [],
...
  "definitions" : {
...
    "json_Result" : {
      "type" : "object",
      "title" : "Result",
          "properties" : {
            "success" : {
            "readOnly" : false,
"description" : "Gives information about the operation.\nIf it is true, operation is successfull\nif it is false, operation is failed",
"type" : "boolean"
            },
            "message" : {
            "readOnly" : false,
"description" : "Message may be filled when error happens so that developer may\nunderstand what the problem is",
"type" : "string"
            }
          },
      "example" : {
  "success" : true,
  "message" : "..."
},
      "description" : ""
    }
...
  },
  "paths": {
    "\/broadcast/addEndpoint" : {
      "post" : {
        "tags" : [ "BroadcastRestService" ],
        "summary" : "Add a third pary rtmp end point to the stream.",
        "description" : "Add a third pary rtmp end point to the stream. When broadcast is started, it will send\nrtmp stream to this rtmp url as well.",
        "operationId" : "addEndpoint",
        "produces" : [ "application/json" ],
        "parameters" : [
          {
            "name" : "id",
            "in" : "formData",
            "type" : "string",
            "description" : "This is the id of broadcast"
          },
          {
            "name" : "rtmpUrl",
            "in" : "formData",
            "type" : "string",
            "description" : "rtmp url of the endpoint that stream will be republished"
          }
        ],
        "responses" : {
          "201" : {
            "schema" : {
"$ref" : "#/definitions/json_Result"
            },
            "example" : {
  "success" : true,
  "message" : "..."
},
            "description" : "io.antmedia.rest.BroadcastRestService.Result"
          }
        }
      }
    }
...
    "\/broadcast/addSocialEndpoint" : {
      "post" : {
        "tags" : [ "BroadcastRestService" ],
        "summary" : "Add social endpoint to a stream.",
        "description" : "Add social endpoint to a stream",
        "operationId" : "addSocialEndpoint",
        "produces" : [ "application/json" ],
        "parameters" : [
          {
            "name" : "id",
            "in" : "formData",
            "type" : "string",
            "description" : "of the broadcast"
          },
          {
            "name" : "serviceName",
            "in" : "formData",
            "type" : "string",
            "description" : "name of the service like facebook, youtube, periscope\nin order to have successfull operation. Social network must be authorized in advance"
          }
        ],
        "responses" : {
          "201" : {
            "schema" : {
"$ref" : "#/definitions/json_Result"
            },
            "example" : {
  "success" : true,
  "message" : "..."
},
            "description" : "io.antmedia.rest.BroadcastRestService.Result"
          }
        }
      }
    }
...
    "\/broadcast/checkDeviceAuthStatus/{serviceName}" : {
      "post" : {
        "tags" : [ "BroadcastRestService" ],
        "summary" : "Check if device is authenticated in the social network.",
        "description" : "Check if device is authenticated in the social network.\nIn authorization phase, this function may be polled periodically until it returns success.\nServer checks social network service for about 1 minute so that if user does not enter DeviceAuthParameters in a 1 minute,\nthis function will never return true",
        "operationId" : "checkDeviceAuthStatus",
        "produces" : [ "application/json" ],
        "parameters" : [
          {
            "name" : "serviceName",
            "in" : "path",
            "required" : true,
            "type" : "string",
            "description" : "Name of the service facebook,youtube,periscope"
          }
        ],
        "responses" : {
          "201" : {
            "schema" : {
"$ref" : "#/definitions/json_Result"
            },
            "example" : {
  "success" : true,
  "message" : "..."
},
            "description" : "Result object with success field.\nIf success field is true, it is authenticated\nif false, not authenticated"
          }
        }
      }
    }
...
    "\/broadcast/create" : {
      "post" : {
        "tags" : [ "BroadcastRestService" ],
        "summary" : "Creates a broadcast and returns the full broadcast object with rtmp address\nand other information.",
        "description" : "Creates a broadcast and returns the full broadcast object with rtmp address\nand other information.",
        "operationId" : "createBroadcast",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [
          {
            "name" : "body",
            "in" : "body",
            "schema" : {
"$ref" : "#/definitions/json_Broadcast"
            },
            "description" : "Broadcast object only related information should be set, it may be null as well."
          }
        ],
        "responses" : {
          "201" : {
            "schema" : {
"$ref" : "#/definitions/json_Broadcast"
            },
            "example" : {
  "date" : 12345,
  "publish" : true,
  "publicStream" : true,
  "category" : "...",
  "status" : "...",
  "streamId" : "...",
  "description" : "...",
  "is360" : true,
  "endPointList" : [ {
    "streamId" : "...",
    "rtmpUrl" : "...",
    "broadcastId" : "...",
    "name" : "...",
    "type" : "..."
  }, {
    "streamId" : "...",
    "rtmpUrl" : "...",
    "broadcastId" : "...",
    "name" : "...",
    "type" : "..."
  } ],
  "duration" : 12345,
  "plannedStartDate" : 12345,
  "name" : "...",
  "listenerHookURL" : "..."
},
            "description" : "io.antmedia.datastore.db.types.Broadcast"
          }
        }
      }
    }
...
    "\/broadcast/createWithSocial" : {
      "post" : {
        "tags" : [ "BroadcastRestService" ],
        "summary" : "Create broadcast and bind social networks at the same time\nServer should be authorized in advance to make this service return success.",
        "description" : "Create broadcast and bind social networks at the same time\nServer should be authorized in advance to make this service return success",
        "operationId" : "createWithSocial",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [
          {
            "name" : "socialNetworks",
            "in" : "query",
            "type" : "string",
            "description" : "Comma separated social network names\nSocial network names must in comma separated and names must match with the defined names\nlike facebook,periscope,youtube etc."
          },
          {
            "name" : "body",
            "in" : "body",
            "schema" : {
"$ref" : "#/definitions/json_Broadcast"
            },
            "description" : "Broadcast\nio.antmedia.datastore.db.types.Broadcast"
          }
        ],
        "responses" : {
          "201" : {
            "schema" : {
"$ref" : "#/definitions/json_Broadcast"
            },
            "example" : {
  "date" : 12345,
  "publish" : true,
  "publicStream" : true,
  "category" : "...",
  "status" : "...",
  "streamId" : "...",
  "description" : "...",
  "is360" : true,
  "endPointList" : [ {
    "streamId" : "...",
    "rtmpUrl" : "...",
    "broadcastId" : "...",
    "name" : "...",
    "type" : "..."
  }, {
    "streamId" : "...",
    "rtmpUrl" : "...",
    "broadcastId" : "...",
    "name" : "...",
    "type" : "..."
  } ],
  "duration" : 12345,
  "plannedStartDate" : 12345,
  "name" : "...",
  "listenerHookURL" : "..."
},
            "description" : "io.antmedia.rest.BroadcastRestService.Result"
          }
        }
      }
    }
...
    "\/broadcast/delete/{id}" : {
      "post" : {
        "tags" : [ "BroadcastRestService" ],
        "summary" : "Delete broadcast from data store\n\nTODO:\nStop publishing if it is being published\nDelete all stream if it is vod\nDelete all stream if vod is stored some under storage.",
        "description" : "Delete broadcast from data store\n\nTODO:\nStop publishing if it is being published\nDelete all stream if it is vod\nDelete all stream if vod is stored some under storage",
        "operationId" : "deleteBroadcast",
        "produces" : [ "application/json" ],
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "required" : true,
            "type" : "string",
            "description" : "Id of the braodcast"
          }
        ],
        "responses" : {
          "201" : {
            "schema" : {
"$ref" : "#/definitions/json_Result"
            },
            "example" : {
  "success" : true,
  "message" : "..."
},
            "description" : "Result object with success field true or false"
          }
        }
      }
    }
...
    "\/broadcast/deleteVoDFile/{id}" : {
      "post" : {
        "tags" : [ "BroadcastRestService" ],
        "summary" : "Deletes vod file in the file system.",
        "description" : "Deletes vod file in the file system",
        "operationId" : "deleteVoDFile",
        "produces" : [ "application/json" ],
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "required" : true,
            "type" : "string",
            "description" : "name of the file"
          }
        ],
        "responses" : {
          "201" : {
            "schema" : {
"$ref" : "#/definitions/json_Result"
            },
            "example" : {
  "success" : true,
  "message" : "..."
},
            "description" : "io.antmedia.rest.BroadcastRestService.Result"
          }
        }
      }
    }
...
    "\/broadcast/get" : {
      "get" : {
        "tags" : [ "BroadcastRestService" ],
        "summary" : "Get broadcast object.",
        "description" : "Get broadcast object",
        "operationId" : "getBroadcast",
        "produces" : [ "application/json" ],
        "parameters" : [
          {
            "name" : "id",
            "in" : "query",
            "type" : "string",
            "description" : "id of the broadcast"
          }
        ],
        "responses" : {
          "200" : {
            "schema" : {
"$ref" : "#/definitions/json_Broadcast"
            },
            "example" : {
  "date" : 12345,
  "publish" : true,
  "publicStream" : true,
  "category" : "...",
  "status" : "...",
  "streamId" : "...",
  "description" : "...",
  "is360" : true,
  "endPointList" : [ {
    "streamId" : "...",
    "rtmpUrl" : "...",
    "broadcastId" : "...",
    "name" : "...",
    "type" : "..."
  }, {
    "streamId" : "...",
    "rtmpUrl" : "...",
    "broadcastId" : "...",
    "name" : "...",
    "type" : "..."
  } ],
  "duration" : 12345,
  "plannedStartDate" : 12345,
  "name" : "...",
  "listenerHookURL" : "..."
},
            "description" : "broadcast object\nnothing if broadcast is not found"
          }
        }
      }
    }
...
    "\/broadcast/getDeviceAuthParameters/{serviceName}" : {
      "post" : {
        "tags" : [ "BroadcastRestService" ],
        "summary" : "Get device parameters for social network authrozation.",
        "description" : "Get device parameters for social network authrozation.",
        "operationId" : "getDeviceAuthParameters",
        "produces" : [ "application/json" ],
        "parameters" : [
          {
            "name" : "serviceName",
            "in" : "path",
            "required" : true,
            "type" : "string",
            "description" : "Name of the service, like facebook,youtube,periscope"
          }
        ],
        "responses" : {
          "201" : {
            "schema" : {
"description" : "If operation is successfull, DeviceAuthParameters is returned with related information.\n\nUser should go to io.antmedia.social.endpoint.VideoServiceEndpoint.DeviceAuthParameters#verification_url\nand enter io.antmedia.social.endpoint.VideoServiceEndpoint.DeviceAuthParameters#user_code in a minute\n\nIf not successfull, it returns with Result object with message",
"type" : "file"
            },
            "description" : "If operation is successfull, DeviceAuthParameters is returned with related information.\n\nUser should go to io.antmedia.social.endpoint.VideoServiceEndpoint.DeviceAuthParameters#verification_url\nand enter io.antmedia.social.endpoint.VideoServiceEndpoint.DeviceAuthParameters#user_code in a minute\n\nIf not successfull, it returns with Result object with message"
          }
        }
      }
    }
...
  }
}